Restructure component and update tests
