+++ src/app/component/ImageUpload/ImageUpload.stories.js
@@ -1,5 +1,5 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
-import ImageUpload from '.';
+import ImageUpload from './ImageUpload';
 
 storiesOf('ImageUpload', module).add('Upload', () => <ImageUpload />);
